chore: create a CEM plugin - defineElementPlugin and use it in both places
